fix compile on unusual configurations
darcs-hash:20060902082344-7aa3a-d01bf2814f41b237f6ddb91e27daf9b26eb6b235

+++ src/SimplePlaybackCore.cxx
@@ -286,7 +286,7 @@ void SimplePlaybackCore::play()
 		m_currentFrame = g_timeline->m_seekPosition;
 		m_lastFrame = g_timeline->m_seekPosition;
 		m_scrubpos = 0;
-		m_scrubmax = (jack_bufsiz!=0)?(int)ceil(scrublen/jack_bufsiz):1;
+		m_scrubmax = (jack_bufsiz!=0)?(int)ceil((double)(scrublen/jack_bufsiz)):1;
 
 		if (!g_use_jack_transport) {
 			m_audioPosition = m_currentFrame * ( 48000 / 25 );
